moved several functions to user service, called logout from search bar, and maintained to stay on page if declining to logout, push to login page after accepting
