+++ src/components/searchbar/searchbar.html
@@ -22,13 +22,13 @@
     <ion-title>My MemePoll</ion-title>
   </ion-toolbar>
   <ion-list id='userSidebar'>
-    <ion-item *ngIf="this.user.loggedIn == false">
+    <ion-item *ngIf="this.userService.loggedIn == false">
       <ion-label (click)='goToLogin()' menuClose>
         Login
       </ion-label>
     </ion-item>
-    <ion-item *ngIf="this.user.loggedIn == true">
-      <ion-label (click)='goToLogout()' menuClose>
+    <ion-item *ngIf="this.userService.loggedIn == true">
+      <ion-label (click)='goToLogout()'>
         Logout
       </ion-label>
     </ion-item>
