+++ src/providers/facebook-oath/facebook-oath.ts
@@ -40,4 +40,6 @@ facebookSDKLoad(){
    }(document, 'script', 'facebook-jssdk'));
 }
 
+
+
 }
